Galprod (#11)
* Create deleteme

* Delete deleteme

* Release (#7)

* Internal info command improvements

* Code organization

* Code organization

* Release (#8)

* Internal info command improvements

* Code organization

* Code organization

* Update appsettings.json

* Release (#9)

* Internal info command improvements

* Code organization

* Code organization

* Minor corrections

* Radio player command improvements

* Release (#10)

* Internal info command improvements

* Code organization

* Code organization

* Minor corrections

* Radio player command improvements

* WIP: added Omada openApi client

* Removed artifact folders

* Updated gitignore

* Delete TeleBotService/OmadaResponses/getClientsResponse.json

Deleted

* Added GetNetClientsCommand and basic Omada open api client

* Added NetClientMonitorCommand

+++ Omada.OpenApi.Client/Responses/PagedResult.cs
@@ -2,7 +2,7 @@
 
 public record PagedResult<TData>
 {
-    public required int TotalRows { get; init; }
+    public required long TotalRows { get; init; }
     public required int CurrentPage { get; init; }
     public required int CurrentSize { get; init; }
     public required IReadOnlyCollection<TData> Data { get; init; }
